Guarantee that absolute paths are returned
As it turns out, some functions are influenced by environment variables
and those variables are not normalized in any way, so we cannot just
parse absolute directories, we should expect that there may be relative
directories as well. We should also canonicalize them when
possible. Solution is to use ‘resolveDir'’ instead of ‘parseAbsDir’.

Functions affected: ‘getHomeDir’ and ‘getTempDir’.